Adding optional parameter
PHP was throwing a warning, I believe, because the second argument was required, but not (always?) passed in. My solution on a friend's site was to just add `= false` there, but the argument is not used anyway, so I'm not sure what good it is doing in there exactly to begin with.