Add purge_dir() and make purge_all() more accurate
purge_all() was previously used to both purge everything in a
directory, and purge an entire site. This adjusts purge_all() to use a
new purge_dir() function when required by environments that use
filesystem for caching, or directly send a wildcard purge for those
that use the nginx layer.