Limit wildcard and non-public post status purges
This prevents purging on posts that did not have a public status either
before or after the save occurred. It switches the hooked action to
transition_post_status instead of save_post to get access to the
previous post status. It also uses the new purge_dir() function instead
of purge_all() for targeting specifically the date archive directory.